requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
python-dotenv==1.0.1
aiohttp==3.9.1
matplotlib==3.8.2